Something interfering with sEMG data fetching. Currently seeking solution

+++ CommandMotorMovement/MovementFunctions.ino
@@ -16,7 +16,7 @@ void Move(int xChange, int yChange, int zChange) {
 
 
 
-void GoTo(int xCoordinate, int yCoordinate, int zCoordinate) { //GoTo for 3 axis(Z included)
+void GoTo(double xCoordinate, double yCoordinate, double zCoordinate) { //GoTo for 3 axis(Z included)
   MoveablePos = {xCoordinate, yCoordinate, zCoordinate};
   //BLA::Matrix<1, 3> MoveablePos = CheckMovementCoordinate(xCoordinate, yCoordinate, zCoordinate); //Checks if the Coordinate positions are reachable
